+++ frontend/src/main.tsx
@@ -7,6 +7,7 @@ import './mobile.css'
 import './polish.css'
 import './mobile-experience.css'
 import './desktop-experience.css'
+import './components/player-themes/vinyl-collection.css'
 
 void registerServiceWorker().catch(() => undefined)